Fix last foot not return when comming from email

+++ src/main/webapp/assets/modules/common/securityService.js
@@ -73,8 +73,11 @@ techlooper.factory("securityService", function (apiService, $rootScope, $q, util
 
       var lastFoot = localStorageService.get("lastFoot");
       if (lastFoot) {
-        localStorageService.remove("lastFoot");
-        return $location.url(lastFoot);
+        var param = $location.search();
+        if (!$.isEmptyObject(param)) {
+          localStorageService.remove("lastFoot");
+          return $location.url(lastFoot);
+        }
       }
 
       switch ($rootScope.userInfo.roleName) {
@@ -111,7 +114,7 @@ techlooper.factory("securityService", function (apiService, $rootScope, $q, util
         .error(function () {utils.sendNotification(jsonValue.notifications.loaded, $(window).height());});
     },
 
-    removeProtectedLastFoot: function() {
+    removeProtectedLastFoot: function () {
       var path = localStorageService.get("protectedPage");
       if (/\/freelancer\/post-project/.test(path) || /\/employer-dashboard/.test(path) || /\/post-challenge/.test(path)) {
         localStorageService.remove("protectedPage");
@@ -149,6 +152,9 @@ techlooper.factory("securityService", function (apiService, $rootScope, $q, util
               return event.preventDefault();
             }
             break;
+
+          default:
+            return localStorageService.set("lastFoot", $location.path());
         }
       });
     }
